+++ client_src/js/printerManagementRunner.js
@@ -784,10 +784,453 @@ bulkPreHeat.addEventListener("click", async (e) => {
 });
 
 let bulkControl = document.getElementById("bulkControl");
-bulkControl.addEventListener("click", async (e) => {});
+bulkControl.addEventListener("click", async (e) => {
+  const printersControl = async function () {
+    let printersToConnect = [];
+    //Grab all check boxes
+    const selectedPrinters = PrinterSelect.getSelected();
+    selectedPrinters.forEach((element) => {
+      const ca = element.id.split("-");
+      printersToConnect.push(ca[1]);
+    });
+    let cameraBlock = ``;
+
+    printersToConnect.forEach((printer) => {
+      const index = _.findIndex(printerInfo, function (o) {
+        return o._id === printer;
+      });
+      if (index > -1) {
+        cameraBlock += `
+        <div class="col-lg-3">
+            <img width="100%" src="${printerInfo[index].cameraURL}">
+        </div>
+        `;
+      }
+    });
+
+    bootbox.dialog({
+      title: "Bulk printer control...",
+      message: `
+      <div id="printerControls" class="row">
+            <div class="col-lg-12">
+              <div class="row">
+                 ${cameraBlock}
+              </div>
+            </div>
+            <div class="col-md-6">
+                            <div class="row">
+                    <div class="col-9">
+                        <center>
+                            <h5>X/Y</h5>
+                        </center>
+                        <hr>
+                    </div>
+                    <div class="col-3">
+                        <center>
+                            <h5>Z</h5>
+                        </center>
+                        <hr>
+                    </div>
+                </div>
+                <div class="row">
+                    <div class="col-3"></div>
+                    <div class="col-3">
+                        <center><button id="pcYpos" type="button" class="btn btn-light"><i class="fas fa-arrow-up"></i></button></center>
+                    </div>
+                    <div class="col-3"></div>
+                    <div class="col-3">
+                        <center><button id="pcZpos" type="button" class="btn btn-light"><i class="fas fa-arrow-up"></i></button></center>
+                    </div>
+                </div>
+                <div class="row">
+                    <div class="col-3">
+                        <center><button id="pcXneg" type="button" class="btn btn-light"><i class="fas fa-arrow-left"></i></button></center>
+                    </div>
+                    <div class="col-3">
+                        <center><button id="pcXYhome" type="button" class="btn btn-light"><i class="fas fa-home"></i></button></center>
+                    </div>
+                    <div class="col-3">
+                        <center><button id="pcXpos" type="button" class="btn btn-light"><i class="fas fa-arrow-right"></i></button></center>
+                    </div>
+                    <div class="col-3">
+                        <center><button id="pcZhome" type="button" class="btn btn-light"><i class="fas fa-home"></i></button></center>
+                    </div>
+                </div>
+                <div class="row">
+                    <div class="col-3"></div>
+                    <div class="col-3">
+                        <center><button id="pcYneg" type="button" class="btn btn-light"><i class="fas fa-arrow-down"></i></button></center>
+                    </div>
+                    <div class="col-3"></div>
+                    <div class="col-3">
+                        <center><button id="pcZneg" type="button" class="btn btn-light"><i class="fas fa-arrow-down"></i></button></center>
+                    </div>
+                </div>
+                <div class="row">
+                    <div class="col-12">
+                        <center>
+                            <div id="pcAxisSteps" class="btn-group" role="group">
+                                <button id="pcAxisSteps01" type="button" class="btn btn-light" value="01">0.1</button>
+                                <button id="pcAxisSteps1" type="button" class="btn btn-light" value="1">1</button>
+                                <button id="pcAxisSteps10" type="button" class="btn btn-dark active" value="10">10</button>
+                                <button id="pcAxisSteps100" type="button" class="btn btn-light" value="100">100</button>
+                            </div>
+                        </center>
+                    </div>
+                </div>
+            </div>
+            <div class="col-md-6">
+            <center>
+            <button id="pmPrintStart" type="button" class="btn btn-success mb-1" role="button" style="display: inline-block;"><i class="fas fa-print"></i> Print</button> <br>
+            <button id="pmPrintPause" type="button" class="btn btn-light mb-1" role="button" style="display: inline-block;"><i class="fas fa-pause"></i> Pause</button> <br>
+            <button id="pmPrintRestart" type="button" class="btn btn-warning mb-1" role="button"  style="display: inline-block;"><i class="fas fa-undo"></i> Restart</button> <br>
+            <button id="pmPrintResume" type="button" class="btn btn-info mb-1" role="button"  style="display: inline-block;"><i class="fas fa-redo"></i> Resume</button> <br>
+            <button id="pmPrintStop" type="button" class="btn btn-danger mb-1" style="display: inline-block;"><i class="fas fa-square"></i> Cancel</button> <br>
+            </center>
+            </div>
+        `,
+      size: "large",
+      onEscape: true,
+      backdrop: true,
+      closeButton: true,
+      onShow: function (e) {
+        //Grab Page
+        const printerControls = {
+          xPlus: document.getElementById("pcXpos"),
+          xMinus: document.getElementById("pcXneg"),
+          yPlus: document.getElementById("pcYpos"),
+          yMinus: document.getElementById("pcYneg"),
+          xyHome: document.getElementById("pcXYhome"),
+          zPlus: document.getElementById("pcZpos"),
+          zMinus: document.getElementById("pcZneg"),
+          zHome: document.getElementById("pcZhome"),
+          step01: document.getElementById("pcAxisSteps01"),
+          step1: document.getElementById("pcAxisSteps1"),
+          step10: document.getElementById("pcAxisSteps10"),
+          step100: document.getElementById("pcAxisSteps100"),
+          printStart: document.getElementById("pmPrintStart"),
+          printPause: document.getElementById("pmPrintPause"),
+          printRestart: document.getElementById("pmPrintRestart"),
+          printResume: document.getElementById("pmPrintResume"),
+          printStop: document.getElementById("pmPrintStop"),
+        };
+        printerControls.printStart.addEventListener("click", async (e) => {
+          printersToConnect.forEach((printer) => {
+            const index = _.findIndex(printerInfo, function (o) {
+              return o._id === printer;
+            });
+            if (index > -1) {
+              e.target.disabled = true;
+              const opts = {
+                command: "start",
+              };
+
+              OctoPrintClient.jobAction(printerInfo[index], opts, e);
+            }
+          });
+        });
+        printerControls.printPause.addEventListener("click", (e) => {
+          printersToConnect.forEach((printer) => {
+            const index = _.findIndex(printerInfo, function (o) {
+              return o._id === printer;
+            });
+            if (index > -1) {
+              e.target.disabled = true;
+              const opts = {
+                command: "pause",
+                action: "pause",
+              };
+              OctoPrintClient.jobAction(printerInfo[index], opts, e);
+            }
+          });
+        });
+        printerControls.printRestart.addEventListener("click", (e) => {
+          printersToConnect.forEach((printer) => {
+            const index = _.findIndex(printerInfo, function (o) {
+              return o._id === printer;
+            });
+            if (index > -1) {
+              e.target.disabled = true;
+              const opts = {
+                command: "restart",
+              };
+              OctoPrintClient.jobAction(printerInfo[index], opts, e);
+            }
+          });
+        });
+        printerControls.printResume.addEventListener("click", (e) => {
+          printersToConnect.forEach((printer) => {
+            const index = _.findIndex(printerInfo, function (o) {
+              return o._id === printer;
+            });
+            if (index > -1) {
+              e.target.disabled = true;
+              const opts = {
+                command: "pause",
+                action: "resume",
+              };
+              OctoPrintClient.jobAction(printerInfo[index], opts, e);
+            }
+          });
+        });
+        printerControls.printStop.addEventListener("click", (e) => {
+          bootbox.confirm({
+            message: `Are you sure you want to cancel all of your ongoing print?`,
+            buttons: {
+              cancel: {
+                label: '<i class="fa fa-times"></i> Cancel',
+              },
+              confirm: {
+                label: '<i class="fa fa-check"></i> Confirm',
+              },
+            },
+            callback(result) {
+              if (result) {
+                printersToConnect.forEach((printer) => {
+                  const index = _.findIndex(printerInfo, function (o) {
+                    return o._id === printer;
+                  });
+                  if (index > -1) {
+                    e.target.disabled = true;
+                    const opts = {
+                      command: "cancel",
+                    };
+                    OctoPrintClient.jobAction(printerInfo[index], opts, e);
+                  }
+                });
+              }
+            },
+          });
+        });
+
+        printerControls.xPlus.addEventListener("click", (e) => {
+          printersToConnect.forEach((printer) => {
+            const index = _.findIndex(printerInfo, function (o) {
+              return o._id === printer;
+            });
+            if (index > -1) {
+              OctoPrintClient.move(e, printerInfo[index], "jog", "x");
+            }
+          });
+        });
+        printerControls.xMinus.addEventListener("click", (e) => {
+          printersToConnect.forEach((printer) => {
+            const index = _.findIndex(printerInfo, function (o) {
+              return o._id === printer;
+            });
+            if (index > -1) {
+              OctoPrintClient.move(e, printerInfo[index], "jog", "x", "-");
+            }
+          });
+        });
+        printerControls.yPlus.addEventListener("click", (e) => {
+          printersToConnect.forEach((printer) => {
+            const index = _.findIndex(printerInfo, function (o) {
+              return o._id === printer;
+            });
+            if (index > -1) {
+              OctoPrintClient.move(e, printerInfo[index], "jog", "y");
+            }
+          });
+        });
+        printerControls.yMinus.addEventListener("click", (e) => {
+          printersToConnect.forEach((printer) => {
+            const index = _.findIndex(printerInfo, function (o) {
+              return o._id === printer;
+            });
+            if (index > -1) {
+              OctoPrintClient.move(e, printerInfo[index], "jog", "y", "-");
+            }
+          });
+        });
+        printerControls.xyHome.addEventListener("click", (e) => {
+          printersToConnect.forEach((printer) => {
+            const index = _.findIndex(printerInfo, function (o) {
+              return o._id === printer;
+            });
+            if (index > -1) {
+              OctoPrintClient.move(e, printerInfo[index], "home", ["x", "y"]);
+            }
+          });
+        });
+        printerControls.zPlus.addEventListener("click", (e) => {
+          printersToConnect.forEach((printer) => {
+            const index = _.findIndex(printerInfo, function (o) {
+              return o._id === printer;
+            });
+            if (index > -1) {
+              OctoPrintClient.move(e, printerInfo[index], "jog", "z");
+            }
+          });
+        });
+        printerControls.zMinus.addEventListener("click", (e) => {
+          printersToConnect.forEach((printer) => {
+            const index = _.findIndex(printerInfo, function (o) {
+              return o._id === printer;
+            });
+            if (index > -1) {
+              OctoPrintClient.move(e, printerInfo[index], "jog", "z", "-");
+            }
+          });
+        });
+        printerControls.zHome.addEventListener("click", (e) => {
+          printersToConnect.forEach((printer) => {
+            const index = _.findIndex(printerInfo, function (o) {
+              return o._id === printer;
+            });
+            if (index > -1) {
+              OctoPrintClient.move(e, printerInfo[index], "home", ["z"]);
+            }
+          });
+        });
+        printerControls.step01.addEventListener("click", (e) => {
+          printersToConnect.forEach((printer) => {
+            const index = _.findIndex(printerInfo, function (o) {
+              return o._id === printer;
+            });
+            if (index > -1) {
+              OctoFarmClient.post("printers/stepChange", {
+                printer: printerInfo[index]._id,
+                newSteps: "01",
+              });
+            }
+          });
+
+          printerControls.step01.className = "btn btn-dark active";
+          printerControls.step1.className = "btn btn-light";
+          printerControls.step10.className = "btn btn-light";
+          printerControls.step100.className = "btn btn-light";
+        });
+        printerControls.step1.addEventListener("click", (e) => {
+          printersToConnect.forEach((printer) => {
+            const index = _.findIndex(printerInfo, function (o) {
+              return o._id === printer;
+            });
+            if (index > -1) {
+              OctoFarmClient.post("printers/stepChange", {
+                printer: printerInfo[index]._id,
+                newSteps: "1",
+              });
+            }
+          });
+
+          printerControls.step1.className = "btn btn-dark active";
+          printerControls.step01.className = "btn btn-light";
+          printerControls.step10.className = "btn btn-light";
+          printerControls.step100.className = "btn btn-light";
+        });
+        printerControls.step10.addEventListener("click", (e) => {
+          printersToConnect.forEach((printer) => {
+            const index = _.findIndex(printerInfo, function (o) {
+              return o._id === printer;
+            });
+            if (index > -1) {
+              OctoFarmClient.post("printers/stepChange", {
+                printer: printerInfo[index]._id,
+                newSteps: "10",
+              });
+            }
+          });
+
+          printerControls.step10.className = "btn btn-dark active";
+          printerControls.step1.className = "btn btn-light";
+          printerControls.step01.className = "btn btn-light";
+          printerControls.step100.className = "btn btn-light";
+        });
+        printerControls.step100.addEventListener("click", (e) => {
+          printersToConnect.forEach((printer) => {
+            const index = _.findIndex(printerInfo, function (o) {
+              return o._id === printer;
+            });
+            if (index > -1) {
+              OctoFarmClient.post("printers/stepChange", {
+                printer: printerInfo[index]._id,
+                newSteps: "100",
+              });
+            }
+          });
+
+          printerControls.step100.className = "btn btn-dark active";
+          printerControls.step1.className = "btn btn-light";
+          printerControls.step10.className = "btn btn-light";
+          printerControls.step01.className = "btn btn-light";
+        });
+      },
+    });
+  };
+
+  PrinterSelect.create(
+    document.getElementById("multiPrintersSection"),
+    false,
+    "Control Printers",
+    printersControl
+  );
+});
 
 let bulkGcodeCommands = document.getElementById("bulkGcodeCommands");
-bulkGcodeCommands.addEventListener("click", async (e) => {});
+bulkGcodeCommands.addEventListener("click", async (e) => {
+  const printersControl = async function () {
+    let printersToConnect = [];
+    //Grab all check boxes
+    const selectedPrinters = PrinterSelect.getSelected();
+    selectedPrinters.forEach((element) => {
+      const ca = element.id.split("-");
+      printersToConnect.push(ca[1]);
+    });
+    bootbox.prompt({
+      size: "small",
+      title: "What gcode commands would you like sent?",
+      inputType: "textarea",
+      callback: function (result) {
+        printersToConnect.forEach(async (printer) => {
+          const index = _.findIndex(printerInfo, function (o) {
+            return o._id === printer;
+          });
+          if (index > -1) {
+            let lines = result.match(/[^\r\n]+/g);
+            lines = lines.map(function (name) {
+              if (!name.includes("=")) {
+                return name.toLocaleUpperCase();
+              } else {
+                return name;
+              }
+            });
+            const opt = {
+              commands: lines,
+            };
+            const post = await OctoPrintClient.post(
+              printerInfo[index],
+              "printer/command",
+              opt
+            );
+            if (post.status === 204) {
+              UI.createAlert(
+                "success",
+                "Your gcode commands have successfully been sent!",
+                3000,
+                "Clicked"
+              );
+            } else {
+              UI.createAlert(
+                "danger",
+                "Your gcode failed to send! Please check the printer is able to receive these commands.",
+                3000,
+                "Clicked"
+              );
+            }
+          }
+        });
+      },
+    });
+  };
+
+  PrinterSelect.create(
+    document.getElementById("multiPrintersSection"),
+    false,
+    "Send Gcode to Printers",
+    printersControl
+  );
+});
 
 function pluginListTemplate(plugin) {
   //Also need check inplace for incompatible...